Try base model with pydantic
