+++ src/ewatercycle/plugins/hype/hype.py
@@ -12,7 +12,7 @@ from dateutil.tz import UTC
 
 from ewatercycle import CFG
 from ewatercycle.container import VersionImages, start_container
-from ewatercycle.models.abstract import AbstractModel
+from ewatercycle.models.abstract import DefaultModel
 from ewatercycle.parameter_sets import ParameterSet
 from ewatercycle.plugins.hype.forcing import HypeForcing
 from ewatercycle.util import geographical_distances, get_time, to_absolute_path
@@ -27,7 +27,7 @@ _version_images: VersionImages = {
 }
 
 
-class Hype(AbstractModel[HypeForcing]):
+class Hype(DefaultModel):
     """eWaterCycle implementation of Hype hydrological model.
 
     Model documentation at http://www.smhi.net/hype/wiki/doku.php .
