+++ src/ewatercycle/plugins/marrmot/marrmot.py
@@ -13,7 +13,7 @@ from cftime import num2date
 
 from ewatercycle import CFG
 from ewatercycle.container import VersionImages, start_container
-from ewatercycle.models.abstract import AbstractModel
+from ewatercycle.models.abstract import DefaultModel
 from ewatercycle.plugins.marrmot.forcing import MarrmotForcing
 from ewatercycle.util import get_time, to_absolute_path
 
@@ -59,7 +59,7 @@ _version_images: VersionImages = {
 }
 
 
-class MarrmotM01(AbstractModel[MarrmotForcing]):
+class MarrmotM01(DefaultModel):
     """eWaterCycle implementation of Marrmot Collie River 1 (traditional bucket) model.
 
     It sets MarrmotM01 parameter with an initial value that is the mean value of
@@ -288,7 +288,7 @@ M14_PARAMS = (
 )
 
 
-class MarrmotM14(AbstractModel[MarrmotForcing]):
+class MarrmotM14(DefaultModel):
     """eWaterCycle implementation of Marrmot Top Model hydrological model.
 
     It sets MarrmotM14 parameter with an initial value that is the mean value of
