+++ src/ewatercycle/plugins/wflow/wflow.py
@@ -12,7 +12,7 @@ from cftime import num2date
 
 from ewatercycle import CFG
 from ewatercycle.container import VersionImages, start_container
-from ewatercycle.models.abstract import AbstractModel
+from ewatercycle.models.abstract import DefaultModel
 from ewatercycle.parameter_sets import ParameterSet
 from ewatercycle.parametersetdb.config import CaseConfigParser
 from ewatercycle.plugins.wflow.forcing import WflowForcing
@@ -40,7 +40,7 @@ _version_images: VersionImages = {
 }
 
 
-class Wflow(AbstractModel[WflowForcing]):
+class Wflow(DefaultModel):
     """Create an instance of the Wflow model class.
 
     Args:
