+++ tests/models/test_abstract.py
@@ -13,7 +13,7 @@ from numpy.testing import assert_array_equal
 
 from ewatercycle import CFG
 from ewatercycle.config import Configuration
-from ewatercycle.models.abstract import AbstractModel
+from ewatercycle.models.abstract import DefaultModel
 from ewatercycle.parameter_sets import ParameterSet
 
 
@@ -28,7 +28,7 @@ def setup_config(tmp_path: Path):
     CFG.reload()
 
 
-class MockedModel(AbstractModel):
+class MockedModel(DefaultModel):
     available_versions = ("0.4.2",)
 
     def __init__(
@@ -88,7 +88,7 @@ def model(bmi: Bmi):
 
 def test_construct():
     with pytest.raises(TypeError) as excinfo:
-        AbstractModel(version="0.4.2")
+        DefaultModel(version="0.4.2")
     msg = str(excinfo.value)
     assert "Can't instantiate abstract class" in msg
     assert "setup" in msg
